Create magazineCode.ino
This is the code for the Arduino located on the magazine. It is a modified version of Samy Kamkar's (@samyk) magspoof which has been converted to be run on arduino by @miaoski. I (@SMR404) added in the Ribbon Sensor functionality and changed the track string to be dynamic, such that it conveys the number of rounds in the magazine.